requirements.txt
python-dotenv==1.0.1
Flask==3.0.1
flask_sqlalchemy==3.1.1
flask_marshmallow==1.1.0
gunicorn==21.2.0
marshmallow-sqlalchemy==0.30.0